requirements.txt
pypdf2
requests
reportlab
pillow